requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.115.14
uvicorn[standard]==0.32.1
pydantic==2.11.7

# Database and SQL dependencies
sqlalchemy==2.0.34
pyodbc==5.2.0
pandas==2.3.0

# Azure services
azure-kusto-data==5.0.4
azure-identity==1.23.0
azure-ai-projects==1.0.0b12
azure-core==1.34.0

# Microsoft Graph API
msgraph-core==1.3.5
msgraph-sdk==1.35.0

# OpenAI integration
openai==1.93.0

# Rate limiting
slowapi==0.1.9

# Environment and configuration
python-dotenv==1.1.1

# SQL parsing and utilities
sqlparse==0.4.4

# Logging
structlog==24.4.0

# HTTP requests
requests==2.32.3

# MCP (Model Context Protocol)
fastmcp==2.9.2

# Report generation
reportlab
xlsxwriter

# Additional utilities
aiohttp
httpx
